Ändrade så att talen sparas som double istället för int. Då kan division bli korrekt även om man inte delar med giltigt heltal.

+++ Inlämningsuppgift1/Program.cs
@@ -9,7 +9,7 @@ namespace Inlämningsuppgift1
         {
 
             List<char> inputOperator = new List<char>();
-            List<int> inputNumber = new List<int>();
+            List<double> inputNumber = new List<double>();
             List<double> sum = new List<double>();
             bool quit = false;
